pyproject.toml
[project]
name = "mailmind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imap-client>=0.1.8",
    "imap-tools>=1.11.0",
    "langchain>=0.3.26",
    "langchain-ollama>=0.3.4",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
]
